- Code refactor for app constant
