+++ composeApp/src/commonMain/kotlin/constant/AppConstant.kt
@@ -1,4 +1,4 @@
-package utils
+package constant
 
 object AppConstant {
     const val BASE_URL = "https://api.themoviedb.org/3/"
