+++ composeApp/src/commonMain/kotlin/ui/tv_series/detail/TvSeriesDetail.kt
@@ -61,7 +61,7 @@ import org.jetbrains.compose.resources.stringResource
 import theme.DefaultBackgroundColor
 import theme.FontColor
 import theme.cornerRadius
-import utils.AppConstant
+import constant.AppConstant
 import utils.roundTo
 
 @Composable
